% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclestats_server.R
\name{cyclestats_server}
\alias{cyclestats_server}
\title{Server for cyclestats app}
\usage{
cyclestats_server(input, output)
}
\value{
\code{server} argument in \code{cyclestatsApp()}
}
\description{
Server for cyclestats app
}
